refactor(migrate): split toSecretId into its own module
Codacy was mis-measuring toSecretId as a 175-line method (it's 10).
Moving it to secret-id.ts side-steps that parser quirk and gives a
cleaner home for the secret-id contract that's also tested directly.
